### 2022-04-01 #4290 @editorialbot
[PRE REVIEW]: glottospace: R package for language mapping and geospatial analysis of linguistic and cultural data
**Submitting author:** @SietzeN (<a href="http://orcid.org/0000-0003-4692-4543">Sietze Norder</a>)
**Repository:** https://github.com/SietzeN/glottospace
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.111
