### 2022-04-01 c63a6ce
references fixed. https://github.com/openjournals/joss-reviews/issues/4290#issuecomment-1085309681
