+++ paper.bib
@@ -17,14 +17,14 @@ journal = {Journal of Statistical Software},
 number = {6},
 pages = {1--39},
 title = {{tmap: Thematic Maps in R}},
-url = {https://doi.org/10.18637/jss.v084.i06},
+url = {10.18637/jss.v084.i06},
 volume = {84},
 year = {2018}
 }
 
 @article{Pebesma2018,
 author = {Pebesma, Edzer},
-doi = {https://doi.org/10.32614/RJ-2018-009},
+doi = {10.32614/RJ-2018-009},
 journal = {The R Journal},
 number = {1},
 pages = {439--446},
