### 2022-04-01 #4290 @SietzeN
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1371/journal.pone.0158391 is OK
> - 10.1038/sdata.2018.205 is OK
> - 10.1007/978-3-540-74686-7 is OK
> - 10.3233/sw-212843 is OK
>
> MISSING DOIs
>
> - 10.18637/jss.v084.i06 may be a valid DOI for title: tmap: Thematic Maps in R
>
> INVALID DOIs
>
> - https://doi.org/10.32614/RJ-2018-009 is INVALID because of 'https://doi.org/' prefix
> ```

References have been fixed in latest commit
