+++ README.md
@@ -10,6 +10,7 @@ stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://
 [![License: GPL
 v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 [![R-CMD-check](https://github.com/SietzeN/glottospace/workflows/R-CMD-check/badge.svg)](https://github.com/SietzeN/glottospace/actions)
+[![status](https://joss.theoj.org/papers/c93a1ccb6835a15c6026ea1ddc28b50c/status.svg)](https://joss.theoj.org/papers/c93a1ccb6835a15c6026ea1ddc28b50c)
 <!-- badges: end -->
 
 # Introduction
@@ -356,11 +357,11 @@ glottosearch(search = "yurakar")
 #> Bounding box:  xmin: -65.1224 ymin: -16.7479 xmax: -65.1224 ymax: -16.7479
 #> Geodetic CRS:  WGS 84
 #>      glottocode     name     macroarea isocode countries family_id
-#> 7543   yura1255 Yuracaré South America     yuz        BO  yura1255
+#> 7546   yura1255 Yuracaré South America     yuz        BO  yura1255
 #>      classification parent_id   family isolate family_size family_size_rank
-#> 7543           <NA>      <NA> Yuracaré    TRUE           1                1
+#> 7546           <NA>      <NA> Yuracaré    TRUE           1                1
 #>      country     continent sovereignty                  geometry
-#> 7543 Bolivia South America     Bolivia POINT (-65.1224 -16.7479)
+#> 7546 Bolivia South America     Bolivia POINT (-65.1224 -16.7479)
 ```
 
 Or limit the search to specific columns:
@@ -373,23 +374,20 @@ glottosearch(search = "Yucuni", columns = c("name", "family"))
 #> Bounding box:  xmin: -97.91818 ymin: -0.76075 xmax: -71.0033 ymax: 17.23743
 #> Geodetic CRS:  WGS 84
 #>      glottocode              name     macroarea isocode countries family_id
-#> 7529   yucu1253            Yucuna South America     ycn  BR;CO;PE  araw1281
-#> 7530   yucu1254 Yucunicoco Mixtec North America                MX  otom1299
+#> 7532   yucu1253            Yucuna South America     ycn  BR;CO;PE  araw1281
+#> 7533   yucu1254 Yucunicoco Mixtec North America                MX  otom1299
 #>                                                      classification parent_id
-#> 7529                            araw1281/japu1236/nucl1764/yucu1252  yucu1252
-#> 7530 otom1299/east2557/amuz1253/mixt1422/mixt1423/mixt1427/sout3179  sout3179
+#> 7532                            araw1281/japu1236/nucl1764/yucu1252  yucu1252
+#> 7533 otom1299/east2557/amuz1253/mixt1422/mixt1423/mixt1427/sout3179  sout3179
 #>           family isolate family_size family_size_rank  country     continent
-#> 7529    Arawakan   FALSE          77               40 Colombia South America
-#> 7530 Otomanguean   FALSE         181               44   Mexico North America
+#> 7532    Arawakan   FALSE          77               40 Colombia South America
+#> 7533 Otomanguean   FALSE         182               44   Mexico North America
 #>      sovereignty                   geometry
-#> 7529    Colombia  POINT (-71.0033 -0.76075)
-#> 7530      Mexico POINT (-97.91818 17.23743)
+#> 7532    Colombia  POINT (-71.0033 -0.76075)
[…]
