### 2022-04-01 c6d96d2
Added WALS codes again. Fixes #73
