+++ man/wals.Rd
@@ -27,4 +27,4 @@ Cite the source of the dataset as:
 Dryer, Matthew S. & Haspelmath, Martin (eds.) 2013. The World Atlas of Language Structures Online. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://wals.info)
 
 
-##------ Fri Apr 01 14:10:44 2022 ------##}
+##------ Fri Apr 01 15:05:36 2022 ------##}
