+++ R/glottodist.R
@@ -1,9 +1,8 @@
-
 #' Calculate distances between languages
 #'
 #' @param glottodata glottodata or glottosubdata, either with or without structure table.
 #' @param structure If glottodata is a table without a structure table, you can add it separately. To create a structure table, you should run glottocreate_structuretable() and you can add it with glottocreate_addtable()
-#' @param id By default, glottodist looks for a column named 'glottocode' or 'glottosubcode', if the id is in a different column, this should be specified.
+#' @param id By default, glottodist looks for a column named 'glottocode', if the id is in a different column, this should be specified.
 #' @return object of class \code{dist}
 #' @export
 #'
@@ -17,7 +16,7 @@ glottodist <- function(glottodata, structure = NULL, id = NULL){
     glottodata <- glottojoin(glottodata)
   } else if(!glottocheck_isglottodata(glottodata)){
     message("glottodata object does not adhere to glottodata/glottosubdata format. Use glottocreate() or glottoconvert().")
-  }
+    }
 
   if(glottocheck_hasmeta(glottodata) & is.null(structure)){
     splitted <- glottosplitmergemeta(glottodata)
@@ -29,6 +28,48 @@ glottodist <- function(glottodata, structure = NULL, id = NULL){
     stop("structure table not found, please add one to glottodata or provide it separately.")
   }
 
+  id <- glottocheck_id(glottodata, id = id)
+
+  duplo <- sum(duplicated(glottodata) | duplicated(glottodata, fromLast = TRUE))
+  if(duplo != 0){
+    message(paste0("This glottodata contains ", duplo, " rows which have at least one duplicate, and ", nrow(glottodata) - duplo, " unique rows \n"))
+    message(paste0("When plotting, you will see ",  nrow(glottodata) - sum(duplicated(glottodata)), " points (unique rows + one of each duplicate) \n"))
+  }
+
+  # structure table:
+  if(!("varname" %in% colnames(structure) ) ){
+    colnames(structure)[1] <- "varname"
+    message("The structure table does not contain a 'varname' column, trying with the first column instead.")
+  }
+
+  glottodata <- tibble::column_to_rownames(glottodata, id)
+
+   if(length(colnames(glottodata)) != length(structure$varname) ){
+    message(paste("The number of variables in ", ifelse(id == "glottocode", "glottodata", "glottosubdata"), "differs from the number of variables in the structure table") )
+    nostruc <- colnames(glottodata)[colnames(glottodata) %nin% structure$varname]
+    novar <- structure$varname[structure$varname %nin% colnames(glottodata)]
+    if(!purrr::is_empty(nostruc)){
+      message(paste0("The following variables exist in the data, but are not defined in the structure table (and will be ignored): ", nostruc))
+    }
+    if(!purrr::is_empty(novar)){
+      message(paste0("The following variables are defined in the structure table but do not exist in the data (and will be ignored): ", novar))
+    }
+  }
+  structure <- suppressMessages(dplyr::left_join(data.frame("varname" = colnames(glottodata)), structure))
+
+  # type
+  if(!("type" %in% colnames(structure) ) ){
+    stop('No type column found in structure. Please add a type column.')
+  }
+
+  dropvars <- which(structure$type %nin% glottocreate_lookuptable()$type_lookup )
+  if(!purrr::is_empty(dropvars)){
+    dropvarnames <- paste0(colnames(glottodata)[dropvars], collapse = ",")
[…]
