+++ R/glottoclean.R
@@ -57,7 +57,9 @@ glottoclean <- function(glottodata, structure = NULL, tona = NULL, tofalse = NUL
                 totrue = all2true,
                 tona = all2na)
 
-  glottodata <- glottoclean_type(glottodata = glottodata, structure = structure, id = id)
+
+
+  # glottodata <- glottoclean_type(glottodata = glottodata, structure = structure, id = id)
 
   if(!is.null(splitted)){
     glottodata <- glottosplitmergemeta(glottodata = glottodata, splitted = splitted)
@@ -76,126 +78,6 @@ glottoclean_all2true <- function(){
 glottoclean_all2na <- function(){
   c("NA", "N A", "N/A", "#N/A", "NA ", " NA", "N /A", "N / A", " N / A", "N / A ", "na", "n a", "n/a",
     "na ", " na", "n /a", "n / a", " a / a", "n / a ", "NULL", "null", "", "\\?", "\\*", "\\.")
-}
-
-#' Set types in glottodata
-#'
-#' @param glottodata glottodata without meta tables
-#' @param structure structure table
-#' @param id By default, glottoclean_type looks for a column named 'glottocode', if the id is in a different column, this should be specified.
-#'
-#' @return a simplified glottodata table (tibble) with column types specified
-#' @noRd
-#'
-#' @examples
-#' glottodata <- glottoget("demodata", meta = TRUE)
-#' glottoclean_type(glottodata, structure = glottodata[["structure"]])
-glottoclean_type <- function(glottodata, structure, id = NULL){
-
-  glottodata <- glottosimplify(glottodata)
-  if(is.null(structure) ){
-    stop("Please provide a structure table.")
-  }
-
-  # Set id:
-  if(is.null(id)){
-    if("glottocode" %in% colnames(glottodata) & "glottosubcode" %nin% colnames(glottodata)){
-      id <- "glottocode"
-      message("glottocode used as id")
-    } else if("glottocode" %nin% colnames(glottodata) & "glottosubcode" %in% colnames(glottodata)){
-      id <- "glottosubcode"
-      message("glottosubcode used as id")
-    } else if(all(c("glottocode", "glottosubcode") %in% colnames(glottodata)) ){
-      id <- "glottocode"
-      message("Data contains glottocodes AND glottosubcodes, glottocode used as id. If this is not what you want, please specify id.")
-    } else if(all(c("glottocode", "glottosubcode") %nin% colnames(glottodata)) ){
-      stop("Please provide an id, or add a 'glottocode' or 'glottosubcode' column to your data")
-    }
-  }
-
-  duplo <- sum(duplicated(glottodata) | duplicated(glottodata, fromLast = TRUE))
-  if(duplo != 0){
-    message(paste0("This glottodata contains ", duplo, " rows which have at least one duplicate, and ", nrow(glottodata) - duplo, " unique rows \n"))
-    message(paste0("When plotting, you will see ",  nrow(glottodata) - sum(duplicated(glottodata)), " points (unique rows + one of each duplicate) \n"))
-  }
-
-  # structure table:
-  if(!("varname" %in% colnames(structure) ) ){
-    colnames(structure)[1] <- "varname"
-    message("The structure table does not contain a 'varname' column, using the first column instead.")
-  }
-
-  rownames(glottodata) <- NULL
-  glottodata <- tibble::column_to_rownames(glottodata, id)
-
-  if(length(colnames(glottodata)) != length(structure$varname) ){
-    message(paste("The number of variables in ", ifelse(id == "glottocode", "glottodata", "glottosubdata"), "differs from the number of variables in the structure table") )
[…]
