+++ man/glottolog.Rd
@@ -27,4 +27,4 @@ Cite the source of the dataset as:
 Hammarström, Harald & Forkel, Robert & Haspelmath, Martin & Bank, Sebastian. 2021. Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)
 
 
-##------ Mon Apr 04 16:54:54 2022 ------##}
+##------ Mon Apr 04 16:58:45 2022 ------##}
