+++ man/dplace.Rd
@@ -23,4 +23,4 @@ This is an internally stored version of D-PLACE data. Use glottoget("dplace", do
 
 This dataset contains the data from D-PLACE, the Database of Places, Language, Culture and Environment, serialized as CLDF StructureDataset.
 
-##------ Mon Apr 04 16:54:54 2022 ------##}
+##------ Mon Apr 04 16:58:46 2022 ------##}
