+++ R/glottodist.R
@@ -118,7 +118,7 @@ glottodist <- function(glottodata, structure = NULL, id = NULL){
   cluster::daisy(x = glottodata, metric = "gower",
                          type = list(symm = symm, asymm = asymm, ordratio = ordratio, logratio = logratio),
                          weights = weights)
-  message("glottodist created")
+
 }
 
 
