+++ man/wals.Rd
@@ -27,4 +27,4 @@ Cite the source of the dataset as:
 Dryer, Matthew S. & Haspelmath, Martin (eds.) 2013. The World Atlas of Language Structures Online. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://wals.info)
 
 
-##------ Mon Apr 04 16:58:46 2022 ------##}
+##------ Mon Apr 04 17:08:14 2022 ------##}
