### 2022-04-05 935a18f
glottoplot: nmds, spotlight, colorby, colorvec
