+++ R/glottoplot.R
@@ -6,8 +6,8 @@
 #' @param type The type of plot: "heatmap", "nmds", or "missing". Default is heatmap if nothing is provided.
 #' @param glottodata glottodata table
 #' @param glottodist A dist object created with \code{\link{glottodist}}
-#' @param color glottovar to be used to color features (optional). Run glottovars() to see the options.
-#' @param label glottovar to be used to label features (optional). Run glottovars() to see the options.
+#' @param colorby Name of variable to be used to color features (optional). Run glottovars() to see the options.
+#' @param label Name of variable to be used to label features (optional). Run glottovars() to see the options.
 #' @param ptsize Size of points between 0 and 1 (optional)
 #' @param filename Optional filename if output should be saved.
 #' @param palette Name of color palette, use glottocolpal("all") to see the options
@@ -17,6 +17,12 @@
 #' @param row2id In case of nmds, specify what each row contains (either 'glottocode' or 'glottosubcode')
 #' @param preventoverlap For nmds with 2 dimensions, should overlap between data points be prevented?
 #' @param alpha For nmds with 2 dimensions: Transparency of points between 0 (very transparent) and 1 (not transparent)
+#' @param colorvec Vector specifying colors for individual values and legend
+#'   order (non-matching values are omitted), for example:
+#'   c("Arawakan" = "rosybrown1", "Yucuna"  = "red",
+#'   "Tucanoan" = "lightskyblue1", "Tanimuca-Retuarã" = "blue", "Naduhup" =
+#'   "gray70", "Kakua-Nukak" = "gray30")
+#'   See the 'values' argument in ggplot2::scale_color_manual() for details.
 #'
 #' @return a visualization of a glotto(sub)data, glottodist or glottonmds object, which can be saved with glottosave()
 #' @export
@@ -27,7 +33,7 @@
 #' glottodata <- glottoget("demodata", meta = TRUE)
 #' glottodist <- glottodist(glottodata = glottodata)
 #' glottoplot(glottodist = glottodist, type = "nmds",
-#'  k = 3, color = "family", label = "name", row2id = "glottocode")
+#'  k = 3, colorby = "family", label = "name", row2id = "glottocode")
 #'
 #' # To create a stress/scree plot, you can run:
 #' # goeveg::dimcheckMDS(matrix = as.matrix(glottodist), k = k)
@@ -39,9 +45,9 @@
 #' glottoplot(glottodata = glottodata, type = "missing")
 #' }
 glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glottonmds = NULL,
-                       color = NULL, ptsize = NULL, label = NULL, filename = NULL, palette = NULL,
+                       colorby = NULL, ptsize = NULL, label = NULL, filename = NULL, palette = NULL,
                        k = NULL, rm.na = FALSE, row2id = NULL,
-                       preventoverlap = FALSE, alpha = NULL){
+                       preventoverlap = FALSE, alpha = NULL, colorvec = NULL){
 
   if(is_dist(glottodata)){
     glottodist <- glottodata
@@ -56,7 +62,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     }
     if(!is.null(glottonmds)){
     glottoplot_nmds(glottonmds = glottonmds,
[…]
