+++ man/glottoget.Rd
@@ -18,14 +18,15 @@ glottoget(
 \item A filepath to locally stored glottodata with file extension (.xlsx .xls .gpkg
 .shp). See also: options meta and simplify.
 \item "glottobase" - Default option, an spatially enhanced version of \href{https://glottolog.org/}{glottolog}. See
-\link{glottobooster} for details.
+\link{glottobooster} for details. If glottodata = NULL, "glottobase" will be loaded.
 \item "wals" - This is a spatially enhanced version of \href{https://wals.info/}{WALS}.
 \item "dplace" - Not yet supported. This is a spatially enhanced version of \href{https://d-place.org/}{D-PLACE}.
 \item "glottolog" - This is a restructured (non-spatial) version of \href{https://glottolog.org/}{glottolog}.
 \item "glottospace" - A simple dataset with glottocodes and a geometry column. This
 is a subset of all languages in \href{https://glottolog.org/}{glottolog} with
 spatial coordinates.
-\item "demodata" - Built-in artificial glottodata (included for demonstration and testing)
+\item "demodata" - Built-in artificial glottodata (included for demonstration and testing).
+\item "demosubdata" - Built-in artificial glottosubdata (included for demonstration and testing)
 }}
 
 \item{meta}{In case 'glottodata' is a path to locally stored data (or demodata/demosubdata): by default, meta sheets are not loaded. Use meta=TRUE if you want to include them.}
