+++ R/glottosplit.R
@@ -44,9 +44,10 @@ glottosplitmergemeta <- function(glottodata, splitted = NULL){
 glottosplit <- function(glottodata){
 
   glottosplit <- vector(mode = "list", length = 2)
-  names(glottosplit) <- c("glottodata", "metadata")
+
 
   if(glottocheck_isglottodata(glottodata)){
+    names(glottosplit) <- c("glottodata", "metadata")
       if(glottocheck_hasmeta(glottodata) ){
         glottosplit[[1]] <- glottodata[["glottodata"]]
         glottosplit[[2]] <- glottodata[names(glottodata) != "glottodata"]
@@ -55,10 +56,16 @@ glottosplit <- function(glottodata){
         glottosplit[[2]] <- NA
       }
   } else if (glottocheck_isglottosubdata(glottodata)){
-    metatables <- names(glottodata) %in% names(glottoget("demodata", meta = TRUE))[-1]
+    names(glottosplit) <- c("glottosubdata", "metadata")
+    metatables <- names(glottodata) %in% names(glottocreate_metadatatables())
     if(glottocheck_hasmeta(glottodata) ){
+      if(sum(!metatables) == 1 & all(names(glottodata)[!metatables] == "glottosubdata")){
+        glottosplit[[1]] <- glottodata[["glottosubdata"]]
+        glottosplit[[2]] <- glottodata[metatables]
+      } else{
         glottosplit[[1]] <- glottodata[!metatables]
         glottosplit[[2]] <- glottodata[metatables]
+      }
     } else {
         glottosplit[[1]] <- glottodata[!metatables]
         glottosplit[[2]] <- NA
