### 2022-04-05 90a24a6
Undo add glottocode to glottosubdata
