### 2022-04-06 3ab9b6c
Some classes added closes #80
