### 2022-04-06 5e8edae
Clean-up glottoclean, remove glottorecode fixes #82. glottosubdata fix glottojoin.
