### 2022-04-06 fa88f02
minor fix glottomap
