+++ R/glottomap.R
@@ -58,7 +58,6 @@ glottomap <- function(glottodata = NULL, color = NULL, label = NULL, type = NULL
       type <- "dynamic"
       message("The country you are trying to plot is too small for a static map, returning a dynamic map instead.")
     }
-
   }
 
   if(!is.null(color) ){
@@ -71,7 +70,10 @@ glottomap <- function(glottodata = NULL, color = NULL, label = NULL, type = NULL
   if(is.null(lbsize) & type == "static"){lbsize <- 0.75}
   if(!is.null(lbsize) & type == "dynamic"){lbsize <- NULL}
 
-  if(!is_sf(glottodata) ) {glottodata <- glottojoin_space(glottodata)}
+  if(!is_sf(glottodata) ) {
+    glottodata <- glottosimplify(glottodata)
+    glottodata <- glottojoin_space(glottodata)
+    }
 
 
 
