+++ R/glottoplot.R
@@ -6,7 +6,7 @@
 #' @param type The type of plot: "heatmap", "nmds", or "missing". Default is heatmap if nothing is provided.
 #' @param glottodata glottodata table
 #' @param glottodist A dist object created with \code{\link{glottodist}}
-#' @param colorby Name of variable to be used to color features (optional). Run glottovars() to see the options.
+#' @param color Name of variable to be used to color features (optional). Run glottovars() to see the options.
 #' @param label Name of variable to be used to label features (optional). Run glottovars() to see the options.
 #' @param ptsize Size of points between 0 and 1 (optional)
 #' @param filename Optional filename if output should be saved.
@@ -33,7 +33,7 @@
 #' glottodata <- glottoget("demodata", meta = TRUE)
 #' glottodist <- glottodist(glottodata = glottodata)
 #' glottoplot(glottodist = glottodist, type = "nmds",
-#'  k = 3, colorby = "family", label = "name", row2id = "glottocode")
+#'  k = 3, color = "family", label = "name", row2id = "glottocode")
 #'
 #' # To create a stress/scree plot, you can run:
 #' # goeveg::dimcheckMDS(matrix = as.matrix(glottodist), k = k)
@@ -45,7 +45,7 @@
 #' glottoplot(glottodata = glottodata, type = "missing")
 #' }
 glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glottonmds = NULL,
-                       colorby = NULL, ptsize = NULL, label = NULL, filename = NULL, palette = NULL,
+                       color = NULL, ptsize = NULL, label = NULL, filename = NULL, palette = NULL,
                        k = NULL, rm.na = FALSE, row2id = NULL,
                        preventoverlap = FALSE, alpha = NULL, colorvec = NULL){
 
@@ -62,7 +62,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     }
     if(!is.null(glottonmds)){
     glottoplot_nmds(glottonmds = glottonmds,
-                      colorby = colorby, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha)
+                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha)
     }
   }
 
@@ -73,7 +73,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     if(type == "nmds" & !is.null(glottodist)){
       glottonmds <- glottonmds(glottodist = glottodist, k = k, rm.na = rm.na, row2id = row2id)
       glottoplot_nmds(glottonmds = glottonmds,
-                      colorby = colorby, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha)
+                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha)
     }
     if(type == "missing"){
       glottoplot_naviewer(data = glottodata, id = "glottocode")
@@ -91,7 +91,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
 #' Plot nmds in 2d or 3d
 #'
[…]
