+++ man/glottomap.Rd
@@ -76,7 +76,7 @@ families <- dplyr::count(glottodata, family, sort = TRUE)
 
 # highlight 10 largest families:
 glottodata <- glottospotlight(glottodata = glottodata, spotcol =
-"family", spotlight = families$family[1:10], spotcontrast = "family", bgcol = "family")
+"family", spotlight = families$family[1:10], spotcontrast = "family", bgcontrast = "family")
 
 # Or, place 10 largest families in background
 glottodata <- glottospotlight(glottodata = glottodata, spotcol =
