+++ R/glottostat.R
@@ -1,0 +1,127 @@
+
+
+#' Permanova across all groups (overall)
+#'
+#' @param glottodata glottodata
+#' @param sample sample table
+#'
+#' @noRd
+#'
+#' @examples
+#' glottodata <- glottoget("demodata", meta = TRUE)
+#' glottostat_permanovall(glottodata)
+glottostat_permanovall <- function(glottodata, sample = NULL, permutations = NULL){
+
+if(is.null(permutations)){permutations <- 1000}
+
+if(glottocheck_hassample(glottodata) & is.null(sample)){
+  glottosample <- glottodata[["sample"]]
+} else if(glottocheck_hassample(glottodata) & !is.null(sample)){
+  glottosample <- sample
+} else if(!glottocheck_hassample(glottodata) & !is.null(sample)){
+  glottosample <- sample
+} else if(!glottocheck_hassample(glottodata) & is.null(sample)){
+  stop("Please provide a sample table")
+}
+
+id <- glottocheck_id(glottodata)
+
+glottodist <- glottodist(glottodata)
+metadist <- glottojoin(glottodata = glottodata, with = glottodist)
+metadist <- glottojoin(glottodata = metadist, with = glottosample)
+
+condist <- metadist %>%
+  dplyr::select(dplyr::all_of(metadist[,id]))
+
+full <- vegan::adonis2(condist ~ group, data = metadist, permutations = permutations)
+p <- round(full[["Pr(>F)"]][1],4)
+
+groupnames <- unique(metadist$group)
+resultsdf <- data.frame(matrix(nrow = 1, ncol = length(groupnames)+2) )
+colnames(resultsdf) <- c(paste0("group", 1:length(groupnames)), "p-value", "significance")
+resultsdf[1,1:length(groupnames)] <- groupnames
+resultsdf[1,"p-value"] <- p
+resultsdf[1,"significance"] <- pvalstars(p)
+
+resultsdf
+
+}
+
+
+
+#' Permanova across all groups (pairwise)
+#'
+#' @param glottodata glottodata
+#' @param sample sample table
+#'
+#' @noRd
+#'
+#' @examples
+#' glottodata <- glottoget("demodata", meta = TRUE)
+#' glottostat_permanovapairs(glottodata)
+glottostat_permanovapairs <- function(glottodata, sample = NULL, permutations = NULL){
+
+  if(is.null(permutations)){permutations <- 1000}
+
+  if(glottocheck_hassample(glottodata) & is.null(sample)){
+    glottosample <- glottodata[["sample"]]
+  } else if(glottocheck_hassample(glottodata) & !is.null(sample)){
+    glottosample <- sample
+  } else if(!glottocheck_hassample(glottodata) & !is.null(sample)){
+    glottosample <- sample
+  } else if(!glottocheck_hassample(glottodata) & is.null(sample)){
+    stop("Please provide a sample table")
+  }
+
+  id <- glottocheck_id(glottodata)
+
+  glottodist <- glottodist(glottodata)
+  metadist <- glottojoin(glottodata = glottodata, with = glottodist)
+  metadist <- glottojoin(glottodata = metadist, with = glottosample)
+
+  # Create empty data.frame to store results
+  groupnames <- unique(metadist$group)
+  resultsdf <- data.frame(matrix(nrow = choose(length(groupnames), 2), ncol = 4) )
+  colnames(resultsdf) <- c("group1", "group2", "p-value", "significance")
[…]
