+++ R/geohelpers.R
@@ -44,22 +44,7 @@ is_lonlat <- function(object){
   sf::st_is_longlat(object)
 }
 
-#' Conditionally transforms dist object to distance matrix
-#'
-#' If dist object is not a distance matrix it will be converted.
-#'
-#' @param dist dist object
-#'
-#' @return distance matrix
-#' @noRd
-#'
-contransform_distmat <- function(dist){
-  if(inherits(dist, what = "dist")){
-    distmat <- as.matrix(dist)
-  } else if (inherits(dist, what = "matrix")){
-    distmat <- dist
-  }
-}
+
 
 #' Check whether crs of two spatial objects are identical
 #'
