### 2022-04-14 91cae3b
demosubdata is now static. Much easier for testing.
