### 2022-04-19 07d0891
join glottosubdata by column, addstructure and addsample
