+++ man/glottocreate_addsample.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/glottocreate.R
+\name{glottocreate_addsample}
+\alias{glottocreate_addsample}
+\title{Add sample table to glottodata or glottosubdata}
+\usage{
+glottocreate_addsample(glottodata)
+}
+\arguments{
+\item{glottodata}{glottodata or glottosubdata}
+}
+\value{
+glottodata/glottosubdata with a sample table
+}
+\description{
+Add sample table to glottodata or glottosubdata
+}
+\examples{
+glottodata <- glottoget("demodata")
+glottocreate_addsample(glottodata)
+}
