### 2022-04-19 7931e32
cmd check notes
