+++ man/grapes-nin-grapes.Rd
@@ -1,0 +1,12 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/glottohelpers.R
+\name{\%nin\%}
+\alias{\%nin\%}
+\title{Negated in}
+\usage{
+a \%nin\% b
+}
+\description{
+Added here to remove dependency on Hmisc
+}
+\keyword{internal}
