+++ NAMESPACE
@@ -1,5 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
+export("%nin%")
 export(glottobooster)
 export(glottoboosterflex)
 export(glottocheck)
@@ -30,5 +31,6 @@ export(glottosimplify)
 export(glottospace)
 export(glottosplitmergemeta)
 export(glottospotlight)
+export(glottostat_permanova)
 importFrom(magrittr,"%>%")
 importFrom(rlang,.data)
