### 2022-04-21 e1974fb
glottocheck: added new checks. Checking metadata is now the default.
