+++ R/TODO.R
@@ -3,7 +3,7 @@
 # library(devtools)
 # document()
 # load_all()
-# install(force = TRUE, args=c("--no-multiarch")) # ALWAYS INSTALL FROM HERE (PACKAGE), NOT FROM OTHER SCRIPT.
+# install(force = TRUE, args=c("--no-multiarch"))
 # knit README
 
 # xfun::gsub_dir(dir = "C:/Users/sjnor/surfdrive/PROJECTS_SN/Rpackages/glottospace/R", pattern = "glottoget_cldf", replacement = "glottoget_cldf")
