### 2022-04-26 a1b051f
colorvec for 3d nmds plots
