### 2022-04-26 e2dd1a3
nmds: fix count of overlapping points
