### 2022-04-26 b0d72c3
fix cmd check
