+++ man/glottomap.Rd
@@ -81,7 +81,7 @@ glottodata <- glottospotlight(glottodata = glottodata, spotcol =
 # Or, place 10 largest families in background
 glottodata <- glottospotlight(glottodata = glottodata, spotcol =
 "family", spotlight = families$family[-c(1:10)], spotcontrast = "family", bgcontrast = "family")
-glottomap(glottodata, color = "color")
+glottomap(glottodata, color = "colpal")
 }
 }
 \concept{<glottomap>}
