### 2022-05-01 SietzeN/glottospace#93 @fauxneticien
JOSS review: Tests for user facing functions
Hi — thanks for creating this package! I got assigned as a JOSS reviewer and am going through the items in the [checklist](https://github.com/openjournals/joss-reviews/issues/4303#issuecomment-1114258877), one of which is:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I see in [`R/TODO.R`](https://github.com/SietzeN/glottospace/blob/main/R/TODO.R#L11) there's this:

```R
# Add test for each user function.
# https://r-pkgs.org/tests.html
```

I was wondering what the status of this to do was, and whether you (all) would like to use the review to get this ball rolling. If so and in case it's helpful, I made a fork of the repository at [fauxneticien/glottospace](https://github.com/fauxneticien/glottospace), and added `test-glottocreate.R`:

```R
test_that("empty glottodata structure is of expected form", {

  empty_df <- glottocreate(glottocodes = c("yucu1253", "tani1257"), variables = 3, meta = FALSE)

  # Is it a data frame?
  expect_true(is.data.frame(empty_df))

  # Is it the expected size?
  expect_equal(ncol(empty_df), 4)
  expect_equal(nrow(empty_df), 2)

  # Other than first column (-1), expect values of all other columns to be NA
  expect_true(all(is.na(unlist(empty_df[, -1], use.names = FALSE))))

})
```

Would it be possible to add at least one (even if relatively simple, to start with) test to each of the 'user function' (also not sure what this subset of functions are)?
