### 2022-05-05 SietzeN/glottospace#93 @SietzeN
p.s. Regarding the user functions:
The majority of functions in the package are not visible to users. There are currently about 20-25 functions that are visible to users. These are the functions that you see in the index of the help pages.
