### 2022-05-09 d4ba3a9
fix cmd check and permanova
