+++ DESCRIPTION
@@ -16,6 +16,7 @@ Depends:
     R (>= 4.1.0)
 Imports: 
     dplyr,
+    plyr,
     ggplot2,
     magrittr,
     purrr,
