### 2022-05-09 d713394
remove NA groups from sample tab
