### 2022-05-10 9a0ad8f
glottocheck is structure table, more tolerant. Now only varname, type, weight columns required.
