### 2022-05-10 809901c
glottoclean: select sample based on sample table
