### 2022-05-13 37ed6f6
Fix issue with naturalearth data, now use geounits and internally stored data instead of ne_countries
