### 2022-05-16 0a01b09
Fix issue sample table
