### 2022-05-16 d5526b4
fix cmd check issues
