+++ R/glottostat.R
@@ -115,7 +115,7 @@ glottostat_permanovapairs <- function(metadist, id, permutations, adj = NULL){
 
   # Select distance matrix
   condist12 <- metadist12 %>%
-    dplyr::select(dplyr::all_of(.data[[id]]))
+    dplyr::select(dplyr::all_of(.[[id]])) # don't use .data
 
   # permutations <- permute::how(nperm = permutations)
   # permute::setBlocks(permutations) <- with(metadist12, glottocode)
