### 2022-05-19 ea34431
drop variables with less than two levels.
