### 2022-05-31 10ff755
glottobooster: rename of columns now optional. In-built glottolog version 4.6
