### 2022-05-31 1875639
glottospace thiessen, use geounits instead of countries. Fixes #95
