### 2022-05-31 17f2867
Do not add glottoclass to sf objects
