### 2022-06-10 #4303 @fauxneticien
@ajstewartlang — I see that there are in-text citations (e.g. `[@Wickham:2016]` on page 3) in the draft paper (https://github.com/openjournals/joss-papers/blob/joss.04303/joss.04303/10.21105.joss.04303.pdf) but the references section is empty. Is that an artefact of the build?
