### 2022-06-10 #4303 @ajstewartlang
👋  @SietzeN can you check the citations in the file paper.md please? The reference list isn't appearing in the pdf build as I believe it's because the citations are not in the appropriate format in paper.md. The backticks aren't need in the markdown file - you can see an example paper with bibliography [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
