### 2022-06-14 SietzeN/glottospace#98 @skalyan91
JOSS review: Software paper
Hi! I’m currently reviewing this package for the Journal of Open Source Software.

One of the requirements for the paper is to “describe how this software compares to other commonly-used packages”. Currently the paper mentions the packages that the glottospace package builds on, but it doesn’t mention other similar packages. This is understandable, as I don’t think there is any other package that does all of what glottospace does! However, I think you should consider mentioning at least the following packages, which also provide access to data from Glottolog and/or WALS:
- [lingtypology](https://docs.ropensci.org/lingtypology/), by George Moroz
- [lingtypR](https://gitlab.com/laurabecker/lingtypr), by Laura Becker
- [qlcData](https://github.com/cysouw/qlcData), by Michael Cysouw

Also, as mentioned in the [review thread](https://github.com/openjournals/joss-reviews/issues/4303), the “References” section is currently empty; see there for further discussion.
