### 2022-06-14 SietzeN/glottospace#99 @skalyan91
JOSS review: Error handling
Hi again,

While trying to run the second example in the readme, I accidentally typed “coronoi” instead of “voronoi”, and spent a few minutes wondering why I was getting dots instead of polygons :). It would be nice if the function gave an error if the `method` argument is anything other than `buffer`, `voronoi`, or `thiessen`.

Not a deal-breaker for the review, but thought it would be good to point out!
