### 2022-06-15 SietzeN/glottospace#100 @skalyan91
JOSS review: Issues with examples in readme
- [x] The example for `glottoclean()` has the line exemplifying the function commented out, and does not show its output.
- [x] The example for `glottojoin()` does not show any output.
- [x] The last line in the example for `glottofilter()` is commented out; it probably shouldn’t be. Also, it would be good to show the output of at least one of the lines.
- [x] In the example for `glottodist()`, the last part (about adding a structure table) seems out of place. I think it should be replaced with just a reference to the `glottocreate()` example.
- [x] #117
- [x] In the second example for `glottosave()` (saving as `.gpkg`), the command should be `glottosave(glottospacedata, filename = "glottospacedata")` (currently it has `glottodata` instead of `glottospacedata`). However, running this command gives me a warning; see [the relevant isssue](https://github.com/SietzeN/glottospace/issues/101).
