### 2022-06-15 SietzeN/glottospace#101 @skalyan91
JOSS review: Saving as .gpkg
When I try to run the following code:
```
glottodata <- glottoget("demodata", meta = FALSE)
glottospacedata <- glottospace(glottodata)
glottosave(glottospacedata, filename = "glottospacedata")
```
I get the following warning:
```
Warning message:
In normalizePath(filename) :
  path[1]="glottospacedata": No such file or directory
```

I’m not sure what this is referring to, as `glottospacedata` is not a filename or directory name.
