### 2022-06-15 SietzeN/glottospace#102 @skalyan91
JOSS review: Instructions for contributors
The review checklist requires “clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support”.

Currently, the “Development” section of the readme covers 2 and 3, but not 1. Please add a brief note to the effect that contributors may submit pull requests on GitHub (if this is how you would like contributions to be made, of course).
