### 2022-06-17 #4303 @skalyan91
Functions to test:

- [ ] glottobooster
- [ ] glottocheck
- [ ] glottoclean
- [ ] glottocode_exists
- [ ] glottoconvert
- [ ] glottocreate
- [ ] glottocreate_addsample
- [ ] glottocreate_addstructure
- [ ] glottodist
- [ ] glottofilter
- [ ] glottofiltermap
- [ ] glottoget
- [ ] glottojoin
- [ ] glottomap
- [ ] glottonmds
- [ ] glottoplot
- [ ] glottosave
- [ ] glottosearch
- [ ] glottosimplify
- [ ] glottospace
- [ ] glottosplitmergemeta
- [ ] glottospotlight
- [ ] glottostat_permanova
