### 2022-06-17 SietzeN/glottospace#103 @skalyan91
JOSS review: glottoclean behaviour
When I run `glottoclean(glottodata)` in the console, I get some messages, but the cleaned-up glottodata is not printed to the console (I can only get this behaviour by running `(glottoclean(glottodata))`). While this is strictly speaking not inconsistent with the documented behaviour of the function, it is unintuitive and may cause confusion as to whether the function is working as expected.

In my opinion, this is not a show-stopper for the review, but I would recommend that it be addressed.
