### 2022-06-17 SietzeN/glottospace#104 @skalyan91
JOSS review: glottoconvert behaviour and documentation
The help file for `glottoconvert()` is missing example code. I would suggest something simple like the following:
```
glottodata <- glottoget("demodata")
glottoconverted <- glottoconvert(glottodata, var = "var")
```

Also, I would recommend (but not require) that `var` be made into an optional argument, with the default value `""` (empty string). My guess is that in many cases, there wouldn’t be a character string that appears in all feature names—or alternatively, even if there is such a string, the user will probably want to keep the feature names as-is for consistency. Assuming that users will want to remove a shared substring seems unintuitive to me.
