### 2022-06-17 SietzeN/glottospace#105 @skalyan91
JOSS review: unimplemented arguments in glottocreate
Currently, the `maintainer`, `email`, `citation`, and `url` arguments of `glottocreate()` do not have any effect. They should be implemented (or else I may be missing something, and would need instructions on how to use them correctly).
