### 2022-06-17 SietzeN/glottospace#106 @skalyan91
JOSS review: use of “regular expression” in glottofilter documentation
The `expression` argument of `glottofilter()` is described as “A regular expression”. This is misleading: I tried using `[aeiou]{2}` to search for all languages with two consecutive vowels in their name, but instead got an error.

Judging by the examples at the end of the help file, what is actually meant is not _regular_ expression, but _logical_ expression. Please fix this.
