### 2022-06-17 SietzeN/glottospace#107 @skalyan91
JOSS review: Instructions for glottomap and glottofiltermap
There needs to be better documentation on how to use interactive selection in `glottomap()` and `glottofiltermap()`. I can’t figure out how to select languages in the selection interface: whenever I click a dot, I get a pop-up with metadata about that language, but I don’t see the selection being saved anywhere, and after I click “Done”, if I evaluate the `selected` object I get `NULL`.
