### 2022-06-17 SietzeN/glottospace#108 @skalyan91
JOSS review: Update D-PLACE support status in glottoget documentation
The documentation for `glottoget()` states that `glottodata = "dplace"` is “Not yet supported”. However, it seems to work (as confirmed also by the fact that [this issue](https://github.com/SietzeN/glottospace/issues/32) is closed). This should be reflected in the documentation.
