### 2022-06-17 SietzeN/glottospace#109 @skalyan91
JOSS review: glottovars() is unimplemented
The description of the `color` and `label` arguments of `glottomap()` (and `glottoplot()`) makes reference to a function `glototvars()`: “Run glottovars() to see glottovars”. It is not clear from this what the functionality of `glottovars()` is supposed to be. In any case it is unimplemented. Either it should be implemented (and documented), or the references to it should be removed.
