### 2022-06-17 SietzeN/glottospace#110 @skalyan91
JOSS review: Projection specification in glottomap
The `'pacific'` value of the `projection` argument of `glottomap()` does not work; I get the following error:
```
Error in `.data[["black"]]`:
! Column `black` not found in `.data`.
Run `rlang::last_error()` to see where the error occurred.
```

In addition, the documentation says that projections other than equal-area and Pacific-centred equal-area may be specified by their “EPSG code”. This is inaccurate: in fact, the function requires the EPSG _description_, not the EPSG code. E.g. for the “[World Mollweide](https://epsg.io/54009)” projection, the EPSG code is 54009, but the description is `+proj=moll +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +datum=WGS84 +units=m no_defs`; only the latter (as a string) is a valid `projection` argument of the `glottomap()` function.
