### 2022-06-17 SietzeN/glottospace#111 @skalyan91
JOSS review: Use na.rm instead of rm.na
Just a recommendation, not a requirement: built-in functions in R such as `mean()` and `sd()` use `na.rm` as the name of the argument that specifies whether to ignore NA values. Using `rm.na` (as is done several times in this package) could be confusing. I would strongly recommend switching to `na.rm` throughout.
