### 2022-06-17 SietzeN/glottospace#112 @skalyan91
JOSS review: Issues with glottonmds
The documentation for the `glottonmds()` function is missing usage examples; I would suggest something like the following:
```
glottodata <- glottoget("demodata", meta = TRUE)
glottodist <- glottodist(glottodata = glottodata)
glottonmds(glottodist, k = 2, row2id = "glottocode")
```

However, this example does not work for me. I get the following output:
```
Run 0 stress 0.02803925
Run 1 stress 0.02803925
... Procrustes: rmse 9.429622e-07  max resid 1.276827e-06
... Similar to previous best
Run 2 stress 0.02803925
... Procrustes: rmse 8.518461e-07  max resid 1.455288e-06
... Similar to previous best
Run 3 stress 0.02803925
... New best solution
... Procrustes: rmse 6.660947e-07  max resid 1.036465e-06
... Similar to previous best
Run 4 stress 0.02803925
... Procrustes: rmse 5.384731e-07  max resid 9.032575e-07
... Similar to previous best
Run 5 stress 0.1230201
Run 6 stress 0.02803925
... Procrustes: rmse 3.60565e-07  max resid 5.639528e-07
... Similar to previous best
Run 7 stress 0.2269841
Run 8 stress 0.1422318
Run 9 stress 0.02803925
... Procrustes: rmse 5.694376e-07  max resid 8.713075e-07
... Similar to previous best
Run 10 stress 0.05028499
Run 11 stress 0.1591543
Run 12 stress 0.1422318
Run 13 stress 0.1230201
Run 14 stress 0.02803925
... Procrustes: rmse 6.493556e-07  max resid 9.9327e-07
... Similar to previous best
Run 15 stress 0.1449615
Run 16 stress 0.02803925
... Procrustes: rmse 3.164524e-07  max resid 4.972362e-07
... Similar to previous best
Run 17 stress 0.05028499
Run 18 stress 0.02803925
... Procrustes: rmse 4.091267e-07  max resid 6.180205e-07
... Similar to previous best
Run 19 stress 0.05028499
Run 20 stress 0.202923
*** Solution reached
Error in x$species[, choices, drop = FALSE] :
  incorrect number of dimensions
```

Furthermore, it is unclear what sort of output this function produces, and how the user is supposed to interact with the output (plot it etc.). The “Value” section in the documentation just says “a glottonmds object”, without specifying what the attributes or available methods of this object are.
