### 2022-06-17 c96406b
fix issue with vegan::scores()
