### 2022-06-17 SietzeN/glottospace#113 @skalyan91
JOSS review: Additional description/references for glottostat_permanova
The `glottostat_permanova()` function seems to work as expected; however, the documentation needs to have some description of what “permanova” is, and ideally have a reference to a publication where the reader can learn more about this statistical method. I don’t think it can be assumed that the average user of this package will be familiar with it.
