### 2022-06-17 #4303 @skalyan91
@ajstewartlang I’ve opened the following issues on the project’s GitHub repository:

1. https://github.com/SietzeN/glottospace/issues/98
2. https://github.com/SietzeN/glottospace/issues/99
3. https://github.com/SietzeN/glottospace/issues/100
4. https://github.com/SietzeN/glottospace/issues/101
5. https://github.com/SietzeN/glottospace/issues/102
6. https://github.com/SietzeN/glottospace/issues/103
7. https://github.com/SietzeN/glottospace/issues/104
8. https://github.com/SietzeN/glottospace/issues/105
9. https://github.com/SietzeN/glottospace/issues/106
10. https://github.com/SietzeN/glottospace/issues/107
11. https://github.com/SietzeN/glottospace/issues/108
12. https://github.com/SietzeN/glottospace/issues/109
13. https://github.com/SietzeN/glottospace/issues/110
14. https://github.com/SietzeN/glottospace/issues/111
15. https://github.com/SietzeN/glottospace/issues/112
16. https://github.com/SietzeN/glottospace/issues/113

Once these issues have been dealt with, I can check off the remaining items on my checklist, at which point my review will be complete.
