+++ man/glottomatch.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/glottomatch.R
+\name{glottomatch}
+\alias{glottomatch}
+\title{glottomatch}
+\usage{
+glottomatch(namevec, glottodata = NULL, tolerance = NULL)
+}
+\arguments{
+\item{namevec}{Vector of language names}
+
+\item{glottodata}{Optional, where to search for matches. If kept empty, the entire glottolog database will be searched, you could also search within a specific area}
+
+\item{tolerance}{Optional, search tolerance.}
+}
+\value{
+
+}
+\description{
+Match a vector of language names to glottocodes and names
+}
+\examples{
+glottodata <- glottofilter(continent = "South America")
+# Finds a single match
+glottomatch(name = "yucuni", glottodata = glottodata)
+# Finds multiple matches
+glottomatch(name = "quechui", glottodata = glottodata)
+}
