+++ NAMESPACE
@@ -23,6 +23,7 @@ export(glottofiltermap)
 export(glottoget)
 export(glottojoin)
 export(glottomap)
+export(glottomatch)
 export(glottonmds)
 export(glottoplot)
 export(glottosave)
