### 2022-07-11 362614a
Glottoconvert fixes #116
