+++ man/wals.Rd
@@ -18,7 +18,7 @@ wals
 This is an internally stored version of WALS data. Use glottoget("wals", download = TRUE) to download the latest version
 }
 \keyword{internal}
-\note{version: v2020.1
+\note{version: v2020.2
 
 
 Cite the source of the dataset as:
@@ -27,4 +27,4 @@ Cite the source of the dataset as:
 Dryer, Matthew S. & Haspelmath, Martin (eds.) 2013. The World Atlas of Language Structures Online. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://wals.info)
 
 
-##------ Fri Jul 01 15:01:53 2022 ------##}
+##------ Mon Jul 11 21:17:31 2022 ------##}
