### 2022-07-15 700bcd4
Fixes #102 JOSS comments, and some minor JOSS issues
