+++ R/glottosave.R
@@ -21,7 +21,7 @@
 #'
 #' glottospacedata <- glottospace(glottodata)
 #' # Saves as .GPKG
-#' glottosave(glottodata, filename = file.path(tempdir(), "glottodata") )
+#' glottosave(glottospacedata, filename = file.path(tempdir(), "glottodata") )
 #'
 #' glottomap <- glottomap(glottodata)
 #' # Saves as .png
