+++ README.md
@@ -40,6 +40,8 @@ confusing. You can help us to improve the package by:
     description of the issue or error message.  
 -   Opening a new issue in the [glottospace issues page on
     GitHub](https://github.com/SietzeN/glottospace/issues)
+-   Fixing a bug or adding functionality and submit a [pull request on
+    GitHub](https://github.com/SietzeN/glottospace/pulls).
 
 # Citation
 
@@ -197,7 +199,8 @@ colnames(glottobase)
 #>  [1] "glottocode"       "name"             "macroarea"        "isocode"         
 #>  [5] "countries"        "family_id"        "classification"   "parent_id"       
 #>  [9] "family"           "isolate"          "family_size"      "family_size_rank"
-#> [13] "country"          "continent"        "sovereignty"      "geometry"
+#> [13] "country"          "sovereignty"      "type"             "geounit"         
+#> [17] "continent"        "adm0_a3"          "geometry"
 ```
 
 ## glottocreate
@@ -227,7 +230,7 @@ summary(glottodata_meta)
 #> references    9     data.frame list
 #> remarks       5     data.frame list
 #> contributors  5     data.frame list
-#> sample        1     data.frame list
+#> sample        3     data.frame list
 #> readme        2     data.frame list
 #> lookup        2     data.frame list
 ```
@@ -239,10 +242,10 @@ added later:
 
 ``` r
 glottocreate_structuretable(varnames = c("var001", "var002", "var003"))
-#>   varname type levels weight groups subgroups
-#> 1  var001   NA     NA      1     NA        NA
-#> 2  var002   NA     NA      1     NA        NA
-#> 3  var003   NA     NA      1     NA        NA
+#>   varname type levels weight group subgroup
+#> 1  var001   NA     NA      1    NA       NA
+#> 2  var002   NA     NA      1    NA       NA
+#> 3  var003   NA     NA      1    NA       NA
 ```
 
 More complex glottodata structures can also be generated. For example,
@@ -268,7 +271,7 @@ glottocreate(glottocodes = c("yucu1253", "tani1257"),
 #> 4 tani1257_b_0002     NA     NA     NA
 #> 
 #> attr(,"class")
-#> [1] "list"          "glottosubdata"
+#> [1] "glottosubdata" "list"
 ```
 
 ## glottocheck
@@ -286,6 +289,7 @@ glottocheck(glottodata, diagnostic = FALSE)
 #> All IDs are valid glottocodes
 #> Some columns have missing data.
 #> Some rows have missing data.
+#> glottodata does not contain metadata
 ```
 
 We’ve now specified diagnostic = FALSE, but the default is to show some
@@ -310,6 +314,8 @@ glottocheck(glottodata, checkmeta = TRUE)
 #> This glottodataset contains the folowing tables: glottodata, structure, description, references, remarks, contributors, sample, readme, lookup
 #> All types recognized
 #> All weights are specified
+#> The following variables are in the data, but there are no such columns variables defined in the structure table: glottocode
+#>  Please check whether the spelling is identical, remove the rows from the structure table, or add the columns to the data.
 ```
 
 <img src="man/figures/README-glottocheckmeta-1.png" width="100%" />
@@ -323,10 +329,10 @@ coding missing values.
 ``` r
 glottodata <- glottoget(glottodata = "demodata", meta = TRUE)
 glottodata$structure
-#>   varname   type levels weight groups subgroups
[…]
