### 2022-07-15 SietzeN/glottospace#112 @SietzeN
The error has been fixed here: https://github.com/SietzeN/glottospace/commit/c96406bbdc370f87dcdd8e45407de9070db6a787#diff-c1f47b46146f6ff49dc17604adeae4cb4277678191e0b0c64497ecdbde897c1eR79
