### 2022-07-15 f441a40
Now use na.rm instead of rm.na throughout the entire package. Closes #111
