### 2022-07-15 dd86a0b
Documentation added for glottostat_permanova. Closes #113
