+++ man/glottostat_permanova.Rd
@@ -21,7 +21,14 @@ glottostat_permanova(
 \item{permutations}{Number of permutations (default is 999)}
 }
 \description{
-Permanova across all groups (overall or pairwise)
+This function takes a glottodata or glottosubdata object and performs a
+Permutational Multivariate Analysis of Variance (PERMANOVA). It can be used
+to test whether two or more groups are significantly different from each
+other (by specifying the 'comparison' argument with either 'overall' or
+'pairwise'). The function uses the 'group' column in the sample table to do
+the comparisons. Before running the analysis, a distance matrix is
+constructed from the glotto(sub)data object using glottodist(). The function
+calls vegan::adonis2(), type ?adonis2 for more details.
 }
 \examples{
 glottodata <- glottoget("demodata", meta = TRUE)
