### 2022-07-15 5dc2f8d
Added example to glottocreate, and clearer description of 'readme' arguments. Closes #105
