### 2022-07-15 SietzeN/glottospace#107 @SietzeN
Hi @skalyan91
This is because the default selection method was set to draw, for which a shape needs to be drawn around the points. I've now changed the default selection mode to 'click'. I also added some lines in the documentation of both functions to make this more explicit. Thanks!
