### 2022-07-15 03da70c
Change default interactive selection mode to 'click' and add some documentation. Closes #107
