### 2022-07-15 SietzeN/glottospace#106 @SietzeN
Absolutely right, thanks for spotting it!
