### 2022-07-15 05b0f6f
glottoconvert. Added clarification as to why 'var' argument is required and added some examples. #104
