### 2022-07-18 bea72ae
glottospace function. Added functionality for water in Voronoi and error handling. Fixes #99 and closes #57.
