### 2022-07-18 117cb4d
Fix message normalizePath, and add option to save as .Rds. Fixes #101
