### 2022-07-18 SietzeN/glottospace#110 @SietzeN
Hi @skalyan91 !

I've fixed the error with the 'pacific' projection.

The function does accept an EPSG code, but this should be the full code, not only the number. I've added an example in the 'projection' parameter description in the glottomap function.

For example: glottomap(projection = "ESRI:54009") plots all languages worldwide in the World Mollweide projection.

I will close this issue with the next push.

Thanks!
Sietze
