### 2022-07-18 a56d075
Projection specification in glottomap. Fixes #110
