### 2022-07-18 8442d5e
glottoconvert JOSS review. Closes #104
