### 2022-07-21 SietzeN/glottospace#100 @SietzeN
Thanks @skalyan91! All six tasks for the read-me have been fixed now.
