### 2022-07-21 2eb442f
Fix issues with readme (#100). Knit readme
