+++ README.md
@@ -236,16 +236,14 @@ summary(glottodata_meta)
 ```
 
 The majority of these meta tables are added for the convenience of the
-user. The ‘structure’ table is the only one that is required for some of
-the functions in the glottospace package. A structure table can also be
-added later:
+user. The ‘structure’ and ‘sample’ tables are the only ones that are
+required for some of the functions in the glottospace package. A
+structure table can also be added later:
 
 ``` r
-glottocreate_structuretable(varnames = c("var001", "var002", "var003"))
-#>   varname type levels weight group subgroup
-#> 1  var001   NA     NA      1    NA       NA
-#> 2  var002   NA     NA      1    NA       NA
-#> 3  var003   NA     NA      1    NA       NA
+glottodata <- glottoget("demodata", meta = FALSE)
+structure <- glottocreate_structuretable(varnames = c("var001", "var002", "var003"))
+glottodata <- glottocreate_addtable(glottodata, structure, name = "structure")
 ```
 
 More complex glottodata structures can also be generated. For example,
@@ -328,15 +326,32 @@ coding missing values.
 
 ``` r
 glottodata <- glottoget(glottodata = "demodata", meta = TRUE)
-glottodata$structure
-#>   varname   type levels weight group subgroup
-#> 1  var001   symm     NA      1    NA       NA
-#> 2  var002 factor     NA      1    NA       NA
-#> 3  var003   symm     NA      1    NA       NA
-# glottodata <- glottoclean(glottodata)
+glottodata_clean <- glottoclean(glottodata)
+#> Values in binary columns (symm/asymm) recoded to TRUE/FALSE
+#> Missing values recoded to NA
+#> All variables have two or more levels (excluding NA)
+#> 
+#>  glottodata has been cleaned.
+
+glottodata$glottodata
+#>   glottocode var001 var002 var003
+#> 1   yucu1253      Y      a      N
+#> 2   tani1257   <NA>      b      Y
+#> 3   ticu1245      Y      a      Y
+#> 4   orej1242      N      b      N
+#> 5   nade1244      N      c      Y
+#> 6   mara1409      N      a      N
+glottodata_clean$glottodata
+#>   glottocode var001 var002 var003
+#> 1   yucu1253   TRUE      a  FALSE
+#> 2   tani1257     NA      b   TRUE
+#> 3   ticu1245   TRUE      a   TRUE
+#> 4   orej1242  FALSE      b  FALSE
+#> 5   nade1244  FALSE      c   TRUE
+#> 6   mara1409  FALSE      a  FALSE
 ```
 
-## glottojoin
+## glottojoin and glottosimplify
 
 Join user-provided glottodata with other datasets, or with online
 databases.
@@ -344,13 +359,56 @@ databases.
 ``` r
 # Join with glottospace
 glottodata <- glottoget("demodata")
-glottodatabase <- glottojoin(glottodata, with = "glottobase")
-glottodataspace <- glottojoin(glottodata, with = "glottospace")
 
-# Join a list of glottodata tables into a single table
-glottodatalist <- glottocreate(glottocodes = c("yucu1253", "tani1257"), 
+# Add data from glottolog:
+glottojoin(glottodata, with = "glottobase")
+#> Simple feature collection with 6 features and 21 fields
+#> Geometry type: POINT
+#> Dimension:     XY
+#> Bounding box:  xmin: -72.4926 ymin: -3.66289 xmax: -66.3068 ymax: -0.59023
+#> Geodetic CRS:  WGS 84
+#>   glottocode var001 var002 var003             name     macroarea isocode
+#> 1   yucu1253      Y      a      N           Yucuna South America     ycn
+#> 2   tani1257   <NA>      b      Y Tanimuca-Retuarã South America     tnc
[…]
