### 2022-07-22 0ed24a6
Read-me add downloads badge and fix some minor issues.
