+++ README.md
@@ -11,6 +11,8 @@ stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://
 v3](https://img.shields.io/badge/License-GPLv3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 [![R-CMD-check](https://github.com/SietzeN/glottospace/workflows/R-CMD-check/badge.svg)](https://github.com/SietzeN/glottospace/actions)
 [![status](https://joss.theoj.org/papers/c93a1ccb6835a15c6026ea1ddc28b50c/status.svg)](https://joss.theoj.org/papers/c93a1ccb6835a15c6026ea1ddc28b50c)
+[![CRAN RStudio mirror
+downloads](https://cranlogs.r-pkg.org/badges/grand-total/glottospace?color=orange)](https://r-pkg.org/pkg/glottospace)
 <!-- badges: end -->
 
 # Introduction
@@ -199,8 +201,7 @@ colnames(glottobase)
 #>  [1] "glottocode"       "name"             "macroarea"        "isocode"         
 #>  [5] "countries"        "family_id"        "classification"   "parent_id"       
 #>  [9] "family"           "isolate"          "family_size"      "family_size_rank"
-#> [13] "country"          "sovereignty"      "type"             "geounit"         
-#> [17] "continent"        "adm0_a3"          "geometry"
+#> [13] "country"          "continent"        "sovereignty"      "geometry"
 ```
 
 ## glottocreate
@@ -230,7 +231,7 @@ summary(glottodata_meta)
 #> references    9     data.frame list
 #> remarks       5     data.frame list
 #> contributors  5     data.frame list
-#> sample        3     data.frame list
+#> sample        1     data.frame list
 #> readme        2     data.frame list
 #> lookup        2     data.frame list
 ```
@@ -269,7 +270,7 @@ glottocreate(glottocodes = c("yucu1253", "tani1257"),
 #> 4 tani1257_b_0002     NA     NA     NA
 #> 
 #> attr(,"class")
-#> [1] "glottosubdata" "list"
+#> [1] "list"          "glottosubdata"
 ```
 
 ## glottocheck
@@ -287,7 +288,6 @@ glottocheck(glottodata, diagnostic = FALSE)
 #> All IDs are valid glottocodes
 #> Some columns have missing data.
 #> Some rows have missing data.
-#> glottodata does not contain metadata
 ```
 
 We’ve now specified diagnostic = FALSE, but the default is to show some
@@ -312,8 +312,6 @@ glottocheck(glottodata, checkmeta = TRUE)
 #> This glottodataset contains the folowing tables: glottodata, structure, description, references, remarks, contributors, sample, readme, lookup
 #> All types recognized
 #> All weights are specified
-#> The following variables are in the data, but there are no such columns variables defined in the structure table: glottocode
-#>  Please check whether the spelling is identical, remove the rows from the structure table, or add the columns to the data.
 ```
 
 <img src="man/figures/README-glottocheckmeta-1.png" width="100%" />
@@ -329,9 +327,6 @@ glottodata <- glottoget(glottodata = "demodata", meta = TRUE)
 glottodata_clean <- glottoclean(glottodata)
 #> Values in binary columns (symm/asymm) recoded to TRUE/FALSE
 #> Missing values recoded to NA
-#> All variables have two or more levels (excluding NA)
-#> 
-#>  glottodata has been cleaned.
 
 glottodata$glottodata
 #>   glottocode var001 var002 var003
@@ -362,7 +357,7 @@ glottodata <- glottoget("demodata")
 
 # Add data from glottolog:
[…]
