+++ paper.md
@@ -1,0 +1,127 @@
+---
+title: 'glottospace: R package for language mapping and geospatial analysis of linguistic and cultural data'
+tags:
+  - language mapping
+  - geolinguistics
+  - linguistic diversity
+  - geographic information systems
+  - comparative linguistics
+  - comparative anthropology
+authors:
+  - name: Sietze Norder
+    orcid: 0000-0003-4692-4543
+    affiliation: "1, 2"
+  - name: Laura Becker
+    orcid: 0000-0002-1835-9404
+    affiliation: 3
+  - name: Hedvig Skirgård
+    orcid: 0000-0002-7748-2381
+    affiliation: 4
+  - name: Leonardo Arias
+    orcid: 0000-0001-6460-7934
+    affiliation: "2, 5"
+  - name: Alena Witzlack-Makarevich
+    orcid: 0000-0003-0138-4635 
+    affiliation: 6
+  - name: Rik van Gijn
+    orcid: 0000-0001-9911-2907
+    affiliation: 2
+affiliations:
+  - name: Copernicus Institute of Sustainable Development, Utrecht University, Utrecht, The Netherlands
+    index: 1
+  - name: Leiden University Centre for Linguistics, Leiden University, Leiden, The Netherlands
+    index: 2
+  - name: Department of General Linguistics, University of Freiburg, Freiburg im Breisgau, Germany
+    index: 3
+  - name: Department of Linguistic and Cultural Evolution, Max-Planck-Institute for Evolutionary Anthropology, Leipzig, Germany
+    index: 4
+  - name: Department of Evolutionary Genetics, Max-Planck-Institute for Evolutionary Anthropology, Leipzig, Germany
+    index: 5
+  - name: Hebrew University of Jerusalem, Jerusalem, Israel
+    index: 6
+date: 30 March 2022
+bibliography: paper.bib
+
+---
+
+# Summary
+
+The glottospace R package facilitates the geospatial analysis of linguistic and cultural data. The aim of the package is to provide a streamlined workflow for geolinguistic analysis, including data entry, data import, cleaning, exploration, language mapping and visualisation, and export. Glottospace is also intended as an R interface to global linguistic and cultural databases such as Glottolog, WALS, and D-PLACE, contributing to improved reproducibility of data analysis.
+
+# Statement of need
+
+Several databases exist that capture aspects of linguistic and cultural diversity globally. However,  there is a lack of interfaces to facilitate the access and manipulation of these databases, specially within the R environment [@R Core Team:2022], or to link these databases with data collected by researchers at smaller scales, or for a particular set of languages. **glottospace** aims to fill this gap by offering a set of functions that automate common tasks in analysing, visualising, and mapping (geo)linguistic data. In essence, the **glottospace** package provides four things: 
+
+1.	Simplified access to global linguistic and cultural databases
+1.	Standardized data structures for data collection, import, cleaning, and checking
+1.	Functionalities to merge data from various linguistic and cultural datasets
+1.	Streamlined workflows for geolinguistic data analysis and visualisation
+
+We will now describe each of these elements in more detail.
+
+Existing global databases of linguistic and cultural diversity, such as Glottolog [@Hammarström:2021], WALS [@Dryer:2013], and D-PLACE [@Kirby:2016], are structured according to the cross-linguistic data format [@Forkel :2018], allowing for the integration of different databases. One way in which languages can be matched across databases is by using glottocodes, i.e. unique identifiers of languages, dialects and language families [@Forkel:2021]. These glottocodes often have geospatial coordinates associated with them, allowing for geospatial analysis and visualisation. With **glottospace**, users can easily access the most recent version of these databases. Researchers can query those databases, use them as a benchmark, or supplement their own data with additional information (like geospatial coordinates, language family).  
+
[…]
