+++ paper.bib
@@ -1,0 +1,133 @@
+@article{Kirby2016,
+doi = {10.1371/journal.pone.0158391},
+issn = {19326203},
+journal = {PLoS ONE},
+number = {7},
+pages = {1--14},
+pmid = {27391016},
+title = {{D-PLACE: A global database of cultural, linguistic and environmental diversity}},
+volume = {11},
+year = {2016}
+}
+
+@article{Tennekes2018,
+author = {Tennekes, M},
+journal = {Journal of Statistical Software},
+number = {6},
+pages = {1--39},
+title = {{tmap: Thematic Maps in R}},
+url = {10.18637/jss.v084.i06},
+volume = {84},
+year = {2018}
+}
+
+@article{Pebesma2018,
+author = {Pebesma, Edzer},
+doi = {10.32614/RJ-2018-009},
+journal = {The R Journal},
+number = {1},
+pages = {439--446},
+title = {{Simple Features for R: Standardized Support for Spatial Vector Data}},
+volume = {10},
+year = {2018}
+}
+
+@book{Dryer2013,
+author = {Dryer, Matthew and Haspelmath, Martin},
+title = {{Dryer, Matthew S. & Haspelmath, Martin (eds.) The World Atlas of Language Structures Online. Leipzig: Max Planck Institute for Evolutionary Anthropology.}},
+year = {2013}
+}
+
+@book{VanGijn2016,
+author = {van Gijn, Rik},
+title = {{Construction-based subordination data (SUB). In Muysken, Pieter et al. (eds.) South American Indian Language Structures (SAILS) Online. Jena: Max Planck Institute for the Science of Human History. (Available at https://sails.clld.org)}},
+year = {2016}
+}
+
+@misc{Bickel2022,
+author = {Bickel, Balthasar and Nichols, Johanna and Zakharko, Taras and {Witzlack-Makarevich, Alena Hildebrandt}, Kristine and Re{\ss}ler, Michael and Bierkandt, Lennart and Z{\'{u}}{\~{n}}iga, Fernando and Lowe, John B},
+title = {{The AUTOTYP database}},
+url = {https://doi.org/10.5281/zenodo.5931509},
+year = {2022}
+}
+
+@misc{Hammarstrom2021,
+author = {Hammarstr{\"{o}}m, Harald and Forkel, Robert and Haspelmath, Martin and Bank, Sebastian},
+title = {{Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)}},
+year = {2021}
+}
+
+@article{Forkel2018,
+author = {Forkel, Robert and List, Johann Mattis and Greenhill, Simon J. and Rzymski, Christoph and Bank, Sebastian and Cysouw, Michael and Hammarstr{\"{o}}m, Harald and Haspelmath, Martin and Kaiping, Gereon A. and Gray, Russell D.},
+doi = {10.1038/sdata.2018.205},
+issn = {20524463},
+journal = {Scientific Data},
+pages = {1--10},
+pmid = {30325347},
+publisher = {The Author(s)},
+title = {{Cross-linguistic data formats, advancing data sharing and re-use in comparative linguistics}},
+volume = {5},
+year = {2018}
+}
+
+@misc{Oksanen2020,
+author = {Oksanen, Jari and Blanchet, F.G and Friendly, M and Kindt, R and Legendre, P and McGlinn, D and Minchin, P.R and O'Hara, R.B and Simpson, G.L and Solymos, P and Stevens, M.H.H and Szoecs, E and Wagner, H},
+title = {{vegan: Community Ecology Package}},
+url = {https://cran.r-project.org/package=vegan},
+year = {2020}
+}
+
+@misc{Wickham2021,
[…]
