+++ paper.md
@@ -50,7 +50,7 @@ The glottospace R package facilitates the geospatial analysis of linguistic and 
 
 # Statement of need
 
-Several databases exist that capture aspects of linguistic and cultural diversity globally. However,  there is a lack of interfaces to facilitate the access and manipulation of these databases, specially within the R environment [@R Core Team:2022], or to link these databases with data collected by researchers at smaller scales, or for a particular set of languages. **glottospace** aims to fill this gap by offering a set of functions that automate common tasks in analysing, visualising, and mapping (geo)linguistic data. In essence, the **glottospace** package provides four things: 
+Several databases exist that capture aspects of linguistic and cultural diversity globally. However, there is a lack of easy-to-use interfaces to access and manipulate these databases, specially within the R environment [@R Core Team:2022], or to link these databases with data collected by researchers at smaller scales or particular set of languages. While other R packages exist for working with linguistic data [@Becker:2022 @Cysouw:2018 @Moroz:2017 @Round:2021]. **glottospace** aims to fill this gap by offering a set of functions that automate common tasks in analysing, visualising, and mapping (geo)linguistic data. In essence, the **glottospace** package provides four things: 
 
 1.	Simplified access to global linguistic and cultural databases
 1.	Standardized data structures for data collection, import, cleaning, and checking
@@ -59,7 +59,7 @@ Several databases exist that capture aspects of linguistic and cultural diversit
 
 We will now describe each of these elements in more detail.
 
-Existing global databases of linguistic and cultural diversity, such as Glottolog [@Hammarström:2021], WALS [@Dryer:2013], and D-PLACE [@Kirby:2016], are structured according to the cross-linguistic data format [@Forkel :2018], allowing for the integration of different databases. One way in which languages can be matched across databases is by using glottocodes, i.e. unique identifiers of languages, dialects and language families [@Forkel:2021]. These glottocodes often have geospatial coordinates associated with them, allowing for geospatial analysis and visualisation. With **glottospace**, users can easily access the most recent version of these databases. Researchers can query those databases, use them as a benchmark, or supplement their own data with additional information (like geospatial coordinates, language family).  
+Existing global databases of linguistic and cultural diversity, such as Glottolog [@Hammarstr{\"{o}}m:2021], WALS [@Dryer:2013], and D-PLACE [@Kirby:2016], are structured according to the cross-linguistic data format [@Forkel:2018], allowing for the integration of different databases. One way in which languages can be matched across databases is by using glottocodes, i.e. unique identifiers of languages, dialects and language families [@Forkel:2021]. These glottocodes often have geospatial coordinates associated with them, allowing for geospatial analysis and visualisation. With **glottospace**, users can easily access the most recent version of these databases. Researchers can query those databases, use them as a benchmark, or supplement their own data with additional information (like geospatial coordinates, language family).  
 
 The **glottospace** package can generate empty data structures to facilitate data entry, or convert existing databases (for example stored in an Excel or CSV file) into two standardized data structures (\autoref{fig:data}): 
 
@@ -80,7 +80,7 @@ The **glottospace** package can generate empty data structures to facilitate dat
 
 ![Examples of glottodata (left) and glottosubdata (right) without metadata tables.\label{fig:data}](joss_data.png)
 
-The glottodata structure is appropriate when one wants to assign one or more features to each language in the dataset (as e.g. in WALS). The glottosubdata structure allows for assigning more complex structures (inventories) to each language in the dataset, which may vary in size from one language to another, such as phoneme inventories (Phoible;  [@Moran:2019] construction/morphological inventories (AUTOTYP; [@Bickel:2022], and subordination strategies in SAILS [@van Gijn:2016]. 
[…]
