+++ paper.bib
@@ -1,4 +1,5 @@
 @article{Kirby:2016,
+author = {Kirby, Kathryn R. and Gray, Russell D. and Greenhill Simon J. and Jordan Fiona M. and Gomes-Ng Stephanie and Bibiko Hans-J{\"{o}}rg and Blasi Damián E. and Botero Carlos A. and Bowern Claire and Ember Carol R. and Leehr Dan and Low Bobbi S. and McCarter Joe and Divale William and Gavin Michael C.},
 doi = {10.1371/journal.pone.0158391},
 issn = {19326203},
 journal = {PLoS ONE},
@@ -38,7 +39,7 @@ title = {{Dryer, Matthew S. & Haspelmath, Martin (eds.) The World Atlas of Langu
 year = {2013}
 }
 
-@book{VanGijn:2016,
+@book{vanGijn:2016,
 author = {van Gijn, Rik},
 title = {{Construction-based subordination data (SUB). In Muysken, Pieter et al. (eds.) South American Indian Language Structures (SAILS) Online. Jena: Max Planck Institute for the Science of Human History. (Available at https://sails.clld.org)}},
 year = {2016}
@@ -51,7 +52,7 @@ url = {https://doi.org/10.5281/zenodo.5931509},
 year = {2022}
 }
 
-@misc{Hammarstrom:2021,
+@misc{Hammarstr{\"{o}}m:2021,
 author = {Hammarstr{\"{o}}m, Harald and Forkel, Robert and Haspelmath, Martin and Bank, Sebastian},
 title = {{Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)}},
 year = {2021}
@@ -130,4 +131,34 @@ publisher = {Springer-Verlag New York},
 title = {{ggplot2: Elegant graphics for data analysis}},
 url = {https://ggplot2.tidyverse.org},
 year = {2016}
+}
+
+@Manual{Round:2021,
+  title = {{glottoTrees}: Phylogenetic trees in Linguistics},
+  author = {Erich R. Round},
+  year = {2021},
+  note = {R package version 0.1},
+  url = {https://github.com/erichround/glottoTrees},
+}
+
+@Manual{Moroz:2017,
+  title = {lingtypology: easy mapping for Linguistic Typology},
+  author = {George Moroz},
+  year = {2017},
+  url = {https://CRAN.R-project.org/package=lingtypology},
+}
+
+@Manual{Becker:2022,
+    title = {lingtypR: Facilitates data processing and illustration in linguistic typology},
+    author = {Laura Becker},
+    year = {2022},
+    note = {R package version 0.0.0.9},
+    url = {https://gitlab.com/laurabecker/lingtypr},
+}
+
+@Manual{Cysouw:2018,
+  title = {qlcData: Processing Data for Quantitative Language Comparison (QLC)},
+  author = {Michael Cysouw},
+  year = {2018},
+  note = {R package version 0.2.1},
 }
