+++ paper.bib
@@ -138,14 +138,14 @@ year = {2016}
   author = {Erich R. Round},
   year = {2021},
   note = {R package version 0.1},
-  url = {https://github.com/erichround/glottoTrees},
+  url = {https://github.com/erichround/glottoTrees}
 }
 
 @Manual{Moroz:2017,
   title = {lingtypology: easy mapping for Linguistic Typology},
   author = {George Moroz},
   year = {2017},
-  url = {https://CRAN.R-project.org/package=lingtypology},
+  url = {https://CRAN.R-project.org/package=lingtypology}
 }
 
 @Manual{Becker:2022,
@@ -153,12 +153,12 @@ year = {2016}
     author = {Laura Becker},
     year = {2022},
     note = {R package version 0.0.0.9},
-    url = {https://gitlab.com/laurabecker/lingtypr},
+    url = {https://gitlab.com/laurabecker/lingtypr}
 }
 
 @Manual{Cysouw:2018,
   title = {qlcData: Processing Data for Quantitative Language Comparison (QLC)},
   author = {Michael Cysouw},
   year = {2018},
-  note = {R package version 0.2.1},
+  note = {R package version 0.2.1}
 }
