+++ paper.bib
@@ -52,7 +52,7 @@ url = {https://doi.org/10.5281/zenodo.5931509},
 year = {2022}
 }
 
-@misc{Hammarstr{\"{o}}m:2021,
+@misc{Hammarström:2021,
 author = {Hammarstr{\"{o}}m, Harald and Forkel, Robert and Haspelmath, Martin and Bank, Sebastian},
 title = {{Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)}},
 year = {2021}
