### 2022-07-22 #4303 @SietzeN
Hi @skalyan91 @fauxneticien and @ajstewartlang,

I have now addressed all comments and fixed all issues. If there are no further issues with the software itself, I will re-submit the markdown file in a new branch (JOSS revision).

Thanks!
Sietze (and co-authors)
