+++ paper.bib
@@ -41,7 +41,8 @@ year = {2013}
 
 @book{vanGijn:2016,
 author = {van Gijn, Rik},
-title = {{Construction-based subordination data (SUB). In Muysken, Pieter et al. (eds.) South American Indian Language Structures (SAILS) Online. Jena: Max Planck Institute for the Science of Human History. (Available at https://sails.clld.org)}},
+title = {{Construction-based subordination data (SUB). In Muysken, Pieter et al. (eds.) South American Indian Language Structures (SAILS) Online. Jena: Max Planck Institute for the Science of Human History}},
+url = {https://sails.clld.org},
 year = {2016}
 }
 
@@ -54,7 +55,8 @@ year = {2022}
 
 @misc{Hammarström:2021,
 author = {Hammarstr{\"{o}}m, Harald and Forkel, Robert and Haspelmath, Martin and Bank, Sebastian},
-title = {{Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)}},
+title = {{Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology,
+url = {https://glottolog.org},
 year = {2021}
 }
 
@@ -149,7 +151,7 @@ year = {2016}
 }
 
 @Manual{Becker:2022,
-    title = {lingtypR: Facilitates data processing and illustration in linguistic typology},
+    title = {lingtypR: easy data manipulation for typologists},
     author = {Laura Becker},
     year = {2022},
     note = {R package version 0.0.0.9},
@@ -161,4 +163,11 @@ year = {2016}
   author = {Michael Cysouw},
   year = {2018},
   note = {R package version 0.2.1}
+}
+
+@misc{Moran:2019,
+author = {Moran S. and McCloy D.},
+title = {PHOIBLE 2.0. Jena: Max Planck Institute for the Science of Human History},
+url = {http://phoible.org},
+year = {2019}
 }
