+++ paper.bib
@@ -166,8 +166,8 @@ year = {2016}
 }
 
 @misc{Moran:2019,
-  author = {Moran S. and McCloy D.},
+  author = {Moran, S. and McCloy, D.},
   title = {PHOIBLE 2.0. Jena: Max Planck Institute for the Science of Human History},
-  url = {http://phoible.org},
-  year = {2019}
+  year = {2019},
+  url = {http://phoible.org}
 }
