+++ paper.bib
@@ -158,16 +158,16 @@ year = {2016}
     url = {https://gitlab.com/laurabecker/lingtypr}
 }
 
-@Manual{Cysouw:2018,
-  title = {qlcData: Processing Data for Quantitative Language Comparison (QLC)},
-  author = {Michael Cysouw},
-  year = {2018},
-  note = {R package version 0.2.1}
-}
-
 @misc{Moran:2019,
   author = {Moran, S. and McCloy, D.},
   title = {PHOIBLE 2.0. Jena: Max Planck Institute for the Science of Human History},
   year = {2019},
   url = {http://phoible.org}
+}
+
+@Manual{Cysouw:2018,
+  title = {qlcData: Processing Data for Quantitative Language Comparison (QLC)},
+  author = {Michael Cysouw},
+  year = {2018},
+  note = {R package version 0.2.1}
 }
