+++ paper.bib
@@ -171,4 +171,3 @@ year = {2016}
   year = {2019},
   url = {http://phoible.org}
 }
-
