+++ paper.bib
@@ -164,10 +164,3 @@ year = {2016}
   year = {2018},
   note = {R package version 0.2.1}
 }
-
-@misc{Moran:2019,
-  author = {Moran, S. and McCloy, D.},
-  title = {{PHOIBLE 2.0. Jena: Max Planck Institute for the Science of Human History}},
-  year = {2019},
-  url = {http://phoible.org}
-}
