+++ paper.bib
@@ -41,8 +41,7 @@ year = {2013}
 
 @book{vanGijn:2016,
 author = {van Gijn, Rik},
-title = {{Construction-based subordination data (SUB). In Muysken, Pieter et al. (eds.) South American Indian Language Structures (SAILS) Online. Jena: Max Planck Institute for the Science of Human History}},
-url = {https://sails.clld.org},
+title = {{Construction-based subordination data (SUB). In Muysken, Pieter et al. (eds.) South American Indian Language Structures (SAILS) Online. Jena: Max Planck Institute for the Science of Human History. (Available at https://sails.clld.org)}},
 year = {2016}
 }
 
@@ -55,8 +54,7 @@ year = {2022}
 
 @misc{Hammarström:2021,
 author = {Hammarstr{\"{o}}m, Harald and Forkel, Robert and Haspelmath, Martin and Bank, Sebastian},
-title = {{Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology,
-url = {https://glottolog.org},
+title = {{Glottolog 4.5. Leipzig: Max Planck Institute for Evolutionary Anthropology. (Available online at https://glottolog.org)}},
 year = {2021}
 }
 
@@ -156,13 +154,6 @@ year = {2016}
     year = {2022},
     note = {R package version 0.0.0.9},
     url = {https://gitlab.com/laurabecker/lingtypr}
-}
-
-@misc{Moran:2019,
-  author = {Moran, S. and McCloy, D.},
-  title = {{PHOIBLE 2.0. Jena: Max Planck Institute for the Science of Human History}},
-  year = {2019},
-  url = {https://phoible.org}
 }
 
 @Manual{Cysouw:2018,
