+++ paper.bib
@@ -162,3 +162,10 @@ year = {2016}
   year = {2018},
   note = {R package version 0.2.1}
 }
+
+@misc{Moran:2019,
+  title = {{PHOIBLE 2.0. Jena: Max Planck Institute for the Science of Human History}},
+  author = {Moran, S. and McCloy, D.},
+  year = {2019},
+  url = {https://phoible.org}
+}
