+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Kirby:2016,
-author = {Kirby, Kathryn R. and Gray, Russell D. and Greenhill Simon J. and Jordan Fiona M. and Gomes-Ng Stephanie and Bibiko Hans-J{\"{o}}rg and Blasi Damián E. and Botero Carlos A. and Bowern Claire and Ember Carol R. and Leehr Dan and Low Bobbi S. and McCarter Joe and Divale William and Gavin Michael C.},
+author = {Kirby, Kathryn R and Gray, Russell D and Greenhill Simon J and Jordan Fiona M and Gomes-Ng Stephanie and Bibiko Hans-J{\"{o}}rg and Blasi Damián E and Botero Carlos A and Bowern Claire and Ember Carol R and Leehr Dan and Low Bobbi S and McCarter Joe and Divale William and Gavin Michael C},
 doi = {10.1371/journal.pone.0158391},
 issn = {19326203},
 journal = {PLoS ONE},
