### 2022-07-25 adeab85
nmds 2 dimensions: label positioning and option to expand grid.
