+++ R/glottoplot.R
@@ -22,6 +22,9 @@
 #'   c("Arawakan" = "rosybrown1", "Yucuna"  = "red",
 #'   "Tucanoan" = "lightskyblue1", "Tanimuca-Retuarã" = "blue", "Naduhup" =
 #'   "gray70", "Kakua-Nukak" = "gray30")
+#' @param expand Optionally expand one or all of the axes. Default is
+#'   c(0,0,0,0), referring to respectively xmin, xmax, ymin, ymax. If you want
+#'   to change the maximum of the x-axis, you would do: c(0,1,0,0).
 #'   See the 'values' argument in ggplot2::scale_color_manual() for details.
 #' @evalRd glottovars()
 #' @return a visualization of a glotto(sub)data, glottodist or glottonmds object, which can be saved with glottosave()
@@ -33,7 +36,7 @@
 #' glottodata <- glottoget("demodata", meta = TRUE)
 #' glottodist <- glottodist(glottodata = glottodata)
 #' # glottoplot(glottodist = glottodist, type = "nmds",
-#' #  k = 3, color = "family", label = "name", row2id = "glottocode")
+#' #  k = 2, color = "family", label = "name", row2id = "glottocode")
 #'
 #' # To create a stress/scree plot, you can run:
 #' # goeveg::dimcheckMDS(matrix = as.matrix(glottodist), k = k)
@@ -47,7 +50,7 @@
 glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glottonmds = NULL,
                        color = NULL, ptsize = NULL, label = NULL, filename = NULL, palette = NULL,
                        k = NULL, na.rm = FALSE, row2id = NULL,
-                       preventoverlap = FALSE, alpha = NULL, colorvec = NULL){
+                       preventoverlap = FALSE, alpha = NULL, colorvec = NULL, expand = NULL){
 
   if(is_dist(glottodata)){
     glottodist <- glottodata
@@ -56,13 +59,15 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     glottodata <- glottodist
   }
 
+  if(!is.numeric(ptsize) & !is.null(ptsize)){stop("ptsize should be a number between 0 and 1")}
+
   if( is.null(type)  ){
     if(is.null(glottonmds)){
     glottoplot_heatmap(glottodist = glottodist, filename = filename)
     }
     if(!is.null(glottonmds)){
     glottoplot_nmds(glottonmds = glottonmds,
-                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec)
+                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand)
     }
   }
 
@@ -73,7 +78,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     if(type == "nmds" & !is.null(glottodist)){
       glottonmds <- glottonmds(glottodist = glottodist, k = k, na.rm = na.rm, row2id = row2id)
       glottoplot_nmds(glottonmds = glottonmds,
-                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec)
[…]
