### 2022-07-27 9795ce1
option label size added for glottoplot_nmds. Closes #120
