+++ R/glottoplot.R
@@ -25,6 +25,7 @@
 #' @param expand Optionally expand one or all of the axes. Default is
 #'   c(0,0,0,0), referring to respectively xmin, xmax, ymin, ymax. If you want
 #'   to change the maximum of the x-axis, you would do: c(0,1,0,0).
+#' @param lbsize Label size (optional)
 #'   See the 'values' argument in ggplot2::scale_color_manual() for details.
 #' @evalRd glottovars()
 #' @return a visualization of a glotto(sub)data, glottodist or glottonmds object, which can be saved with glottosave()
@@ -50,7 +51,7 @@
 glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glottonmds = NULL,
                        color = NULL, ptsize = NULL, label = NULL, filename = NULL, palette = NULL,
                        k = NULL, na.rm = FALSE, row2id = NULL,
-                       preventoverlap = FALSE, alpha = NULL, colorvec = NULL, expand = NULL){
+                       preventoverlap = FALSE, alpha = NULL, colorvec = NULL, expand = NULL, lbsize = NULL){
 
   if(is_dist(glottodata)){
     glottodist <- glottodata
@@ -67,7 +68,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     }
     if(!is.null(glottonmds)){
     glottoplot_nmds(glottonmds = glottonmds,
-                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand)
+                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand, lbsize = lbsize)
     }
   }
 
@@ -78,7 +79,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     if(type == "nmds" & !is.null(glottodist)){
       glottonmds <- glottonmds(glottodist = glottodist, k = k, na.rm = na.rm, row2id = row2id)
       glottoplot_nmds(glottonmds = glottonmds,
-                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand)
+                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand, lbsize = lbsize)
     }
     if(type == "missing"){
       glottoplot_naviewer(data = glottodata, id = "glottocode")
@@ -112,6 +113,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
 #' @param expand Optionally expand one or all of the axes. Default is
 #'   c(0,0,0,0), referring to respectively xmin, xmax, ymin, ymax. If you want
 #'   to change the maximum of the x-axis, you would do: c(0,1,0,0).
+#' @param lbsize Label size (optional)
 #'
 #' @noRd
 #'
@@ -122,13 +124,13 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
 #'
[…]
