### 2022-08-01 SietzeN/glottospace#103 @skalyan91
@SietzeN I think this is a good solution. Thanks!
