### 2022-08-01 SietzeN/glottospace#98 @skalyan91
@SietzeN The paper at https://github.com/openjournals/joss-papers/blob/joss.04303/joss.04303/10.21105.joss.04303.pdf (which I recompiled just now) is still missing the references, and does not mention the packages I pointed out. Am I looking in the wrong place?
