### 2022-08-01 #4303 @skalyan91
@ajstewartlang @SietzeN I’ve tested the `joss_revision` branch of the package, and am satisfied that all of the issues I raised have been addressed. However, I am not able to see the revised version of the paper; I just asked editorialbot to recompile the paper, but it’s still showing the same version that I saw before (no references, no mention of existing packages with similar functionality). Once this final issue is resolved, my review will be complete.
