### 2022-08-02 SietzeN/glottospace#98 @SietzeN
Hi @skalyan91!

Thanks for your mail! That's strange, I've implemented your suggestions and recompiled the paper to test it. The markdown file is here: https://github.com/SietzeN/glottospace/blob/joss_revision/paper.md I'm not sure why it doesn't work. I'll also reply on the JOSS GitHub.
Thanks!
p.s. I'll be away for two weeks, so my reply might take a bit longer.
