### 2022-08-02 #4303 @SietzeN
Hi @skalyan91 and @ajstewartlang

Thanks for your mail! That's strange, I've implemented your suggestions and recompiled the paper to test it. The markdown file is here: https://github.com/SietzeN/glottospace/blob/joss_revision/paper.md I'm not sure why it doesn't work. Perhaps I've made a mistake, should I place the md file somewhere else?
Thanks!
