### 2022-08-08 #4303 @ajstewartlang
:wave:  @SietzeN I just did a PR on one small thing in the paper. If you could merge that and then do the following please, that would be great:

- Make a [tagged release](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository) of the software.
