+++ R/glottoplot.R
@@ -26,6 +26,8 @@
 #'   c(0,0,0,0), referring to respectively xmin, xmax, ymin, ymax. If you want
 #'   to change the maximum of the x-axis, you would do: c(0,1,0,0).
 #' @param lbsize Label size (optional)
+#' @param ptshift (optional) If preventoverlap is TRUE, how much should points be shifted?
+#' @param lbshift (optional) If preventoverlap is TRUE, how much should labels be shifted?
 #'   See the 'values' argument in ggplot2::scale_color_manual() for details.
 #' @evalRd glottovars()
 #' @return a visualization of a glotto(sub)data, glottodist or glottonmds object, which can be saved with glottosave()
@@ -51,7 +53,7 @@
 glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glottonmds = NULL,
                        color = NULL, ptsize = NULL, label = NULL, filename = NULL, palette = NULL,
                        k = NULL, na.rm = FALSE, row2id = NULL,
-                       preventoverlap = FALSE, alpha = NULL, colorvec = NULL, expand = NULL, lbsize = NULL){
+                       preventoverlap = FALSE, alpha = NULL, colorvec = NULL, expand = NULL, lbsize = NULL, ptshift = NULL, lbshift = NULL){
 
   if(is_dist(glottodata)){
     glottodist <- glottodata
@@ -68,7 +70,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     }
     if(!is.null(glottonmds)){
     glottoplot_nmds(glottonmds = glottonmds,
-                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand, lbsize = lbsize)
+                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand, lbsize = lbsize, ptshift = NULL, lbshift = NULL)
     }
   }
 
@@ -79,7 +81,7 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
     if(type == "nmds" & !is.null(glottodist)){
       glottonmds <- glottonmds(glottodist = glottodist, k = k, na.rm = na.rm, row2id = row2id)
       glottoplot_nmds(glottonmds = glottonmds,
-                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand, lbsize = lbsize)
+                      color = color, ptsize = ptsize, label = label, palette = palette, filename = filename, preventoverlap = preventoverlap, alpha = alpha, colorvec = colorvec, expand = expand, lbsize = lbsize, ptshift = NULL, lbshift = NULL)
     }
     if(type == "missing"){
       glottoplot_naviewer(data = glottodata, id = "glottocode")
@@ -114,6 +116,8 @@ glottoplot <- function(glottodata = NULL, glottodist = NULL, type = NULL, glotto
 #'   c(0,0,0,0), referring to respectively xmin, xmax, ymin, ymax. If you want
 #'   to change the maximum of the x-axis, you would do: c(0,1,0,0).
 #' @param lbsize Label size (optional)
[…]
