### 2022-08-30 47349c7
Select sample in glottoclean
