### 2022-08-30 1a4affe
Data cleaning messages added: number of variables dropped in each step
