+++ paper.md
@@ -39,7 +39,7 @@ affiliations:
     index: 5
   - name: Hebrew University of Jerusalem, Jerusalem, Israel
     index: 6
-date: 22 July 2022
+date: 09 September 2022
 bibliography: paper.bib
 
 ---
