### 2022-09-09 #4303 @SietzeN
Dear @ajstewartlang,

Thanks again for editing and managing our submission.

We've created a tagged release of the software: https://github.com/SietzeN/glottospace/releases/tag/v0.0.113

And archived the version here: 10.5281/zenodo.7064529

All the best!
Sietze
