### 2022-09-11 #4303 @danielskatz
Hi @SietzeN - I'm the AEiC who will do the final processing of your submission. I'm suggesting a few small changes in https://github.com/SietzeN/glottospace/pull/123. Please merge this, or let me know what you disagree with, and we can move to acceptance and publication.
