### 2024-01-04 #4303 @ajstewartlang
:wave: @SietzeN I am handling a new JOSS submission that seems to overlap nicely with your expertise. Might you have the bandwidth to review it please? It would be very much appreciated. The submission can be found here: https://github.com/marccanby/linguiphyr/
