### 2024-02-16 #4303 @SietzeN
Hi Andrew,

Sorry for my late reply. Are you still looking for a reviewer? I’ll be away next week but could pick it up after my return.

Cheers,
Sietze

Van: Andrew Stewart ***@***.***>
Verzonden: donderdag 4 januari 2024 17:35
Aan: openjournals/joss-reviews ***@***.***>
CC: Norder, S.J. (Sietze) ***@***.***>; Mention ***@***.***>
Onderwerp: Re: [openjournals/joss-reviews] [REVIEW]: glottospace: R package for language mapping and geospatial analysis of linguistic and cultural data (Issue #4303)

CAUTION: This email originated from outside of Utrecht University. Do not click links or open attachments unless you recognize the sender and know the content is safe.

👋 @SietzeN<https://github.com/SietzeN> I am handling a new JOSS submission that seems to overlap nicely with your expertise. Might you have the bandwidth to review it please? It would be very much appreciated. The submission can be found here: https://github.com/marccanby/linguiphyr/
